python/requirements.txt
pytest
paramiko
allure-pytest  # Added Allure integration